Title: 재귀함수를 이용한 3N + 1 수열, Time: 7ms, Memory: 0MB, Status: Passed - Codetree

+++ 251227/재귀함수를 이용한 3N + 1 수열/3n-plus-1-sequence-with-recursive-function.cpp
@@ -1,7 +1,7 @@
 #include <iostream>
 using namespace std;
 
-long long recur_seq(int num);
+long long recur_seq(long long num);
 
 int main() 
 {
